add:feature from kiro

+++ server/controllers/PostsControllers.js
@@ -4,7 +4,7 @@ import Comments from "../models/CommentsModels.js";
 export const getPosts = async (req, res) => {
   try {
     const post = await Posts.find()
-      .select("author image body  comments")
+      .select("userId author image body comments")
       .populate({
         path: "comments",
         select: "body author",
@@ -67,11 +67,9 @@ export const deletePosts = async (req, res) => {
   try {
     const post = await Posts.findById(req.params.id);
     if (!post) return res.status(404).json({ msg: "Post not found" });
-    // if (req.role != "admin") {
-    //   if (post.userId.toString() != req.userId) {
-    //     return res.sendStatus(403);
-    //   }
-    // }
+    if (post.userId.toString() != req.userId) {
+      return res.sendStatus(403);
+    }
 
     await Posts.findByIdAndDelete(req.params.id);
     await Comments.deleteMany({ postId: req.params.id });
